requirements.txt
streamlit
google-cloud-dialogflow-cx
google-auth
fpdf
requests
